Mouse: Move mouse files to mouse/ subdir
