+++ xbmc/input/InputManager.h
@@ -34,8 +34,8 @@
 #include "Action.h"
 #include "windowing/XBMC_events.h"
 #include "input/mouse/interfaces/IMouseInputProvider.h"
+#include "input/mouse/MouseStat.h"
 #include "input/KeyboardStat.h"
-#include "input/MouseStat.h"
 #include "interfaces/IActionListener.h"
 #include "settings/lib/ISettingCallback.h"
 #include "threads/CriticalSection.h"
