+++ xbmc/input/mouse/MouseStat.cpp
