+++ xbmc/input/mouse/MouseTranslator.h
