+++ xbmc/windowing/osx/WinEventsSDL.cpp
@@ -27,9 +27,9 @@
 #include "GUIUserMessages.h"
 #include "settings/DisplaySettings.h"
 #include "guilib/GUIWindowManager.h"
+#include "input/mouse/MouseStat.h"
 #include "input/Key.h"
 #include "input/InputManager.h"
-#include "input/MouseStat.h"
 #include "windowing/WinSystem.h"
 #include "platform/darwin/osx/CocoaInterface.h"
 #include "ServiceBroker.h"
