+++ xbmc/windowing/win10/WinEventsWin10.cpp
@@ -20,11 +20,11 @@
 
 #include "Application.h"
 #include "guilib/GUIWindowManager.h"
-#include "input/Action.h"
-#include "input/ActionIDs.h"
-#include "input/MouseStat.h"
+#include "input/mouse/MouseStat.h"
 #include "input/touch/generic/GenericTouchActionHandler.h"
 #include "input/touch/generic/GenericTouchInputHandler.h"
+#include "input/Action.h"
+#include "input/ActionIDs.h"
 #include "messaging/ApplicationMessenger.h"
 #include "rendering/dx/DeviceResources.h"
 #include "rendering/dx/RenderContext.h"
